+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handle new video response for upcoming videos in the Local API.
 - Handle new streams tab response in the Local API.
 - Try to avoid SABR-only player responses when using Local API.
+- Handle the case where YouTube returns stream formats without URLs.
 
 ## [0.9.0] - 2026-04-28
 ### Added